+++ src/main/java/lotto/controller/LottoController.java
@@ -1,6 +1,16 @@
-package lotto;
+package lotto.controller;
 
 import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.LottoGenerator;
+import lotto.domain.Lottos;
+import lotto.domain.Money;
+import lotto.domain.vo.Number;
+import lotto.domain.Rank;
+import lotto.domain.Store;
+import lotto.domain.WinnerLotto;
+import lotto.view.InputView;
+import lotto.view.OutputView;
 
 public class LottoController {
