+++ src/main/java/lotto/controller/Rate.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.controller;
 
 import java.math.BigDecimal;
 import java.util.Objects;
