+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,10 +1,11 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import lotto.domain.vo.Number;
 
 public class LottoGenerator {
