+++ src/main/java/lotto/view/InputView.java
@@ -1,11 +1,11 @@
-package lotto;
+package lotto.view;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.Scanner;
 import java.util.stream.Collectors;
+import lotto.domain.vo.Number;
 
 public class InputView {
