+++ src/main/java/lotto/view/OutputView.java
@@ -1,8 +1,12 @@
-package lotto;
+package lotto.view;
 
 import java.text.MessageFormat;
 import java.util.List;
 import java.util.stream.Collectors;
+import lotto.domain.Lotto;
+import lotto.domain.vo.Number;
+import lotto.domain.Rank;
+import lotto.controller.Rate;
 
 public class OutputView {
