+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,7 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 import java.math.BigDecimal;
+import lotto.controller.Rate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
