+++ src/test/java/lotto/domain/WinnerLottoTest.java
@@ -1,11 +1,12 @@
-package lotto;
+package lotto.domain;
 
+import static org.assertj.core.api.Assertions.*;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import lotto.domain.vo.Number;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import static org.assertj.core.api.Assertions.*;
 
 public class WinnerLottoTest {
     private static final Number BONUS = new Number(7);
