+++ src/test/java/lotto/domain/vo/NumberTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.vo;
 
 import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
